+++ README.md
@@ -10,6 +10,6 @@ iex -S mix
 
 Then, start the code (in iex shell)
 ```
-iex(1)> AeChannelRunner.start_channel_helper()
+iex(1)> ChannelRunner.start_channel_helper()
 ```
 requires OTP 20.3
