+++ apps/ae_socket_connector/.gitignore
@@ -20,5 +20,4 @@ erl_crash.dump
 *.ez
 
 # Ignore package tarball (built via "mix hex.build").
-aesocketconnector-*.tar
-
+socket_connector-*.tar
