+++ apps/ae_socket_connector/README.md
@@ -1,21 +1,21 @@
-# AeSocketConnector
+# SocketConnector
 
 **TODO: Add description**
 
 ## Installation
 
 If [available in Hex](https://hex.pm/docs/publish), the package can be installed
-by adding `aesocketconnector` to your list of dependencies in `mix.exs`:
+by adding `socket_connector` to your list of dependencies in `mix.exs`:
 
 ```elixir
 def deps do
   [
-    {:aesocketconnector, "~> 0.1.0"}
+    {:socket_connector, "~> 0.1.0"}
   ]
 end
 ```
 
 Documentation can be generated with [ExDoc](https://github.com/elixir-lang/ex_doc)
 and published on [HexDocs](https://hexdocs.pm). Once published, the docs can
-be found at [https://hexdocs.pm/aesocketconnector](https://hexdocs.pm/aesocketconnector).
+be found at [https://hexdocs.pm/socket_connector](https://hexdocs.pm/socket_connector).
