+++ apps/ae_socket_connector/test/aesocketconnector_test.exs
@@ -1,0 +1,8 @@
+defmodule SocketConnectorTest do
+  use ExUnit.Case
+  doctest SocketConnector
+
+  test "greets the world" do
+    assert SocketConnector.hello() == :world
+  end
+end
