--- apps/aesocketconnector/lib/aetestaccounts.ex
@@ -1,25 +1,0 @@
-defmodule AeTestAccounts do
-  def responderPubkey() do
-    <<145, 57, 82, 197, 159, 203, 87, 93, 38, 245, 163, 158, 237, 249, 101, 141, 158, 185, 198,
-      87, 190, 11, 15, 96, 80, 225, 138, 111, 252, 37, 59, 79>>
-  end
-
-  def responderPrivkey() do
-    <<55, 112, 8, 133, 136, 186, 103, 209, 225, 173, 157, 98, 179, 248, 227, 75, 64, 253, 175, 97,
-      81, 149, 27, 108, 35, 160, 80, 16, 121, 176, 159, 138, 145, 57, 82, 197, 159, 203, 87, 93,
-      38, 245, 163, 158, 237, 249, 101, 141, 158, 185, 198, 87, 190, 11, 15, 96, 80, 225, 138,
-      111, 252, 37, 59, 78>>
-  end
-
-  def initiatorPubkey do
-    <<106, 184, 29, 213, 77, 73, 184, 77, 59, 65, 33, 156, 241, 78, 239, 173, 39, 2, 126, 254,
-      111, 28, 73, 150, 6, 150, 66, 20, 47, 81, 213, 154>>
-  end
-
-  def initiatorPrivkey() do
-    <<133, 143, 10, 3, 177, 135, 2, 205, 204, 153, 181, 19, 83, 137, 93, 186, 100, 92, 12, 201,
-      228, 174, 194, 70, 27, 220, 3, 227, 212, 32, 203, 247, 106, 164, 29, 213, 77, 73, 184, 77,
-      59, 65, 33, 156, 241, 78, 239, 173, 39, 2, 126, 254, 111, 28, 73, 150, 6, 150, 66, 20, 47,
-      81, 213, 153>>
-  end
-end
