--- apps/aesocketconnector/test/aesocketconnector_test.exs
@@ -1,8 +1,0 @@
-defmodule AeSocketConnectorTest do
-  use ExUnit.Case
-  doctest AeSocketConnector
-
-  test "greets the world" do
-    assert AeSocketConnector.hello() == :world
-  end
-end
